Global SASS variables, new distribution boarding page
