+++ src/components/AddressInputs/SelectAddress/SelectAddress.scss
@@ -21,7 +21,7 @@
 
     .selectAddress_info_title {
       font-family: Georgia, 'Times New Roman', Times, serif;
-      color: #2A5EC7;
+      color: var(--primary-color);
       font-size: 18px;
       font-weight: 500;
       margin-bottom: 5px;
@@ -29,7 +29,7 @@
 
     .selectAddress_info_subtitle {
       font-family: Arial, Helvetica, sans-serif;
-      color: black;
+      color: var(--black-color);
       font-size: 12px;
       font-weight: 400;
     }
@@ -49,10 +49,10 @@
     width: 30px;
     height: 30px;
     border-radius: 100%;
-    border: 1px solid #2A5EC7;
+    border: 1px solid var(--primary-color);
     background: #f7f5f5;
     font-size: 12px;
-    color: #2A5EC7;
+    color: var(--primary-color);
     margin-right: 5px;
   }
 }
