+++ src/components/common/CustomButton/CustomButton.scss
@@ -2,15 +2,15 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  background: #2A5EC7;
+  background: var(--primary-color);
   border: none;
   border-radius: 4px;
-  color: white;
+  color: var(--white-color);
+  font-family: var(--secondary-font);
   font-size: 12px;
+  font-weight: 100;
   margin: 10px auto;
   min-height: 40px;
-  font-family: arial;
-  font-size: 11px;
   padding: 0 20px;
 
 
@@ -19,7 +19,7 @@
   }
   
   &:hover {
-    background: #2A5EC7;
+    background: var(--primary-color);
     cursor: pointer;
     box-shadow: 1px 3px 5px rgba(0, 0, 0, 0.2);
   }
