+++ src/components/common/ServicesCard/ServicesCard.scss
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  background: white;
+  background: var(--white-color);
   border: 1px solid #e5e5e5;
   max-width: 200px;
   border-radius: 5px;
@@ -38,7 +38,7 @@
     display: table;
     font-size: 60px;
     font-weight: 600;
-    color: #2A5EC7;
+    color: var(--primary-color);
 
     span {
       font-size: 15px;
@@ -57,14 +57,14 @@
     width: 25px;
     height: 25px;
     margin: 20px 0 0;
-    background-color: white;
+    background-color: var(--white-color);
     border: 1px solid grey;
 
     &:checked {
-      border: 1px solid #2A5EC7;
+      border: 1px solid var(--primary-color);
       padding: 9px;
-      background-color: #2A5EC7;
-      color: white;
+      background-color: var(--primary-color);
+      color: var(--white-color);
     }
   }
 }
