+++ src/components/common/TextInput/TextInput.scss
@@ -2,7 +2,8 @@
   position: relative;
   display: flex;
   flex-direction: column;
-  border: 1px solid black;
+  font-family: var(--secondary-font);
+  border: 1px solid var(--black-color);
   border-radius: 5px;
   height: 40px;
   margin: 10px;
@@ -13,18 +14,18 @@
     background: none;
     width: 100%;
     height: 100%;
-    color: #2A5EC7;
+    color: var(--primary-color);
     font-size: 13px;
     padding: 0 8px;
 
     &::placeholder {
-      color: grey;
-      font-size: 11px;
-      font-weight: bold;
+      color: var(--grey-light-color);
+      font-size: 12px;
+      font-weight: 600;
     }
 
     &:focus::-webkit-input-placeholder {
-      color: white;
+      color: var(--white-color);
     }
 
     &:focus {
@@ -33,7 +34,7 @@
 
     &:focus + label {
       display: block;
-      color: black;
+      color: var(--black-color);
       top: -8px;
       left: 10px;
       font-size: 10px;
@@ -48,10 +49,10 @@
     display: block;
     position: absolute;
     font-size: 12px;
-    color: white;
+    color: var(--white-color);
     top: 12px;
     left: 8px;
-    background: white;
+    background: var(--white-color);
     padding: 0 2px;
     font-weight: bold;
     opacity: 0;
@@ -59,7 +60,7 @@
 
   .keepLabel {
     display: block;
-    color: black;
+    color: var(--black-color);
     top: -8px;
     left: 10px;
     font-size: 10px;
