+++ src/pages/Boarding/sections/Section1st/index.js
@@ -1,0 +1,1 @@
+export { default as Section1st } from './Section1st';
