+++ src/pages/Boarding/sections/Section2nd/index.js
@@ -1,0 +1,1 @@
+export { default as Section2nd } from './Section2nd';
