+++ src/pages/Boarding/sections/Section4th/index.js
@@ -1,0 +1,1 @@
+export { default as Section4th } from './Section4th';
