+++ src/pages/Boarding/sections/Section6th/Section6th.jsx
@@ -1,0 +1,18 @@
+import { Text } from "@/components/common/Text"
+
+import '../../Boarding.scss'
+
+const Section6th = () => {
+  return (
+    <>
+      <Text
+        classStyle='bigTitle'
+      >Thank you for trusting Honest Network.</Text>
+      <Text
+        classStyle='alternativeTitle'
+      >Your data is now in our database,<br/> we will be in touch soon.</Text>
+    </>
+  )
+}
+
+export default Section6th
