+++ src/pages/Boarding/sections/Section6th/index.js
@@ -1,0 +1,1 @@
+export { default as Section6th } from './Section6th';
