--- src/sass/abstract/variables.scss
@@ -1,4 +1,0 @@
-$primary-color: "#094EB8";
-$white-color: "#ffffff";
-$black-color: "#000000";
-$grey-color: "#A3A5A9"
